+++ lib/flat_segmented_control.dart
@@ -4,103 +4,70 @@ import 'package:flutter/material.dart';
 
 class FlatSegmentedControl extends StatefulWidget {
   final List<Widget> children;
-  final List<Widget> navChildren;
-  final EdgeInsetsGeometry padding;
-  final EdgeInsetsGeometry navPadding;
-  final double height;
+  final List<Widget> tabChildren;
+  final double tabHeight;
+  final Color indicatorColor;
+  final Color labelColor;
+  final Color unselectedLabelColor;
 
   FlatSegmentedControl({
     Key key,
     @required this.children,
-    @required this.navChildren,
-    @required this.height,
-    this.padding = const EdgeInsets.all(0.0),
-    this.navPadding = const EdgeInsets.all(0.0),
-  }) : super(key: key);
+    @required this.tabChildren,
+    this.tabHeight = 52.0,
+    this.indicatorColor,
+    this.labelColor,
+    this.unselectedLabelColor,
+  })  : assert(children != null),
+        assert(tabChildren != null),
+        assert(tabChildren.length == children.length),
+        super(key: key);
 
+  @override
   _FlatSegmentedControlState createState() => _FlatSegmentedControlState();
 }
 
-class _FlatSegmentedControlState extends State<FlatSegmentedControl> {
-  int _selected = 0;
-  final _pageController = PageController();
+class _FlatSegmentedControlState extends State<FlatSegmentedControl>
+    with SingleTickerProviderStateMixin {
+  TabController _tabController;
+
+  @override
+  void initState() {
+    super.initState();
+    _tabController = TabController(
+      vsync: this,
+      length: widget.children.length,
+    );
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _tabController?.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Column(
+    return Flex(
+      direction: Axis.vertical,
       children: <Widget>[
-        Container(
-          padding: widget.navPadding,
-          child: Row(
-            children: _getListNavs(),
-          ),
+        TabBar(
+          controller: _tabController,
+          tabs: widget.tabChildren,
+          indicatorColor:
+              widget.indicatorColor ?? Theme.of(context).primaryColor,
+          labelColor:
+              widget.labelColor ?? Theme.of(context).textTheme.body1.color,
+          unselectedLabelColor: widget.unselectedLabelColor ??
+              Theme.of(context).textTheme.body1.color,
         ),
-        Container(
-          padding: widget.padding,
-          height: widget.height,
-          child: PageView(
-            physics: NeverScrollableScrollPhysics(),
-            controller: _pageController,
+        Expanded(
+          child: TabBarView(
+            controller: _tabController,
             children: widget.children,
           ),
         ),
       ],
-    );
-  }
-
-  List<Widget> _getListNavs() {
-    return List<Widget>.generate(
-      widget.navChildren.length,
-      (index) => NavButton(
-        selected: _selected,
-        index: index,
-        navChildren: widget.navChildren,
-        onPressed: () {
-          setState(() {
-            _selected = index;
-            _pageController.animateToPage(
-              index,
-              curve: Curves.easeInOut,
-              duration: Duration(milliseconds: 400),
-            );
-          });
-        },
-      ),
-    );
-  }
-}
-
-class NavButton extends StatelessWidget {
-  final int selected;
-  final int index;
-  final VoidCallback onPressed;
-  final List<Widget> navChildren;
-
-  NavButton({
-    @required this.selected,
-    @required this.index,
-    @required this.onPressed,
-    @required this.navChildren,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        border: Border(
-          bottom: BorderSide(
-            width: 2.5,
-            color: selected == index
-                ? Theme.of(context).primaryColor
-                : Theme.of(context).canvasColor,
-          ),
-        ),
-      ),
-      child: FlatButton(
-        materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-        onPressed: onPressed,
-        child: navChildren[index],
-      ),
     );
   }
 }