+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## [0.0.1] - TODO: Add release date.
 
 * TODO: Describe initial release.
+
+## [0.0.2] - Remove Flexible and replaced with height.
+
+* Now required to insert height of children.
+
+## [0.1.0] - Now based on TabBar and TabBarView.
+
+* Constructor has also been changed. Read README.md for more info.